// src/types/book.ts
export type Book = {
    id: number;
    title: string;
    description: string;
    price: number;
    discount: number;
    rating: number;
    is_best_seller: boolean;
    cover: string;
    year: number;
    edition: string;
    stock: number;
    sales: number;
    isbn: string;
    author_id: number;
    genre_id: number;
    publisher_id: number;
    create_at: string;
    update_at: string;
};